added ft_putnbr, ft_putstr, ft_putchar, ft_strmap, ft_strmapi(not sure if right) also modified make to compile main(reminder: remove before submitting)

+++ ft_strnstr.c
@@ -14,14 +14,14 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int i;
+	size_t i;
 
 	if (needle[0] == '\0')
 		return ((char *)haystack);
 	while (*haystack != '\0')
 	{
 		i = 0;
-		while (haystack[i] == needle[i] && (size_t)i < len)
+		while (haystack[i] == needle[i] && i < len)
 		{
 			if (needle[i + 1] == '\0')
 				return ((char *)haystack);
